+++ CHANGELOG.md
@@ -431,3 +431,4 @@
 - 2024-06-13: Init bettercontact, email opener, mod,  outreach base, +5
 - 2024-06-14: Test  outreach base, summaries, leads, contacts, +4
 - 2024-06-14: Impl mailbox, email opener, mutate queries, embeddings, +2 query
+- 2024-06-15: Pagination contacts, leads, user, send, +4